+++ v3/crates/compatibility/src/compatibility_date.rs
@@ -168,7 +168,8 @@ pub fn get_compatibility_date_for_flag(flag: Flag) -> Option<CompatibilityDate> 
         }
         Flag::RequireValidCommandOutputType => Some(new_compatibility_date(2025, 2, 27)),
         Flag::ValidateObjectTypeDataConnectorTypeMappingFieldTypes
-        | Flag::ValidateArgumentMappingTypes => {
+        | Flag::ValidateArgumentMappingTypes
+        | Flag::DisallowInvalidHeadersInAuthConfig => {
             //TODO: Should be set on next release
             None
         }